renamed files, made exportable mock
