+++ engine/BUILD
@@ -1,7 +1,10 @@
 cc_library(
     name = "engine",
-    srcs = ["matching.cc"],
-    hdrs = ["matching.h"],
+    srcs = ["matcher.cc"],
+    hdrs = [
+        "base_matcher.h",
+        "matcher.h",
+    ],
     visibility = [
         "//entry:__pkg__",
     ],
@@ -12,15 +15,15 @@ cc_library(
 )
 
 cc_library(
-    name = "mock-matching",
-    hdrs = ["mock-matching.h"],
+    name = "mock-matcher",
+    hdrs = ["mock_matcher.h"],
     visibility = ["//entry:__pkg__"],
     deps = ["//engine"],
 )
 
 cc_test(
-    name = "matching-test",
-    srcs = ["matching_test.cc"],
+    name = "matcher-test",
+    srcs = ["matcher_test.cc"],
     deps = [
         "//common",
         "//engine",
