--- engine/base_matching.h
@@ -1,20 +1,0 @@
-//
-// Base class of matcher for use by testers
-//
-
-#pragma once
-/*
-#include <vector>
-
-using std::vector;
-
-class BaseMatcher {
-    virtual ~BaseMatcher() = default;
-    vector<MatchedTrade>, get_matched_trades, (const vector<Trade>& trades_in), (override));
-    PNL, get_pnl_from, (Trade trade), (override));
-    PNL, get_net_pnl, (const vector<Trade>& trades_in, Timestamp end_time = now()), (override));
-    YearEndPNL, get_year_end_pnl, (const vector<Trade>& trades_in, Timestamp year = now()), (override));
-    vector<SnapshotPNL>, get_pnl_snapshots, (const vector<Trade>& trades_in, vector<TimeDelta> timestamps = DEFAULT_SAMPLES), (override));
-    vector<Trade>, get_earliest_long_term_sells, (const vector<Trade>& trades), (override));
-};
-*/
